# takeout-server/API-Swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "NodeJS Server for Take-Out App.",
    "version": "1.0.0",
    "title": "Take Out Server"
  },
  "host": "take-out-server.herokuapp.com",
  "schemes": [
    "https"
  ],
...
  "paths": {
...
    "/menus/{id}": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Get specified menu with dishes",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful",
            "schema": {
              "$ref": "#/definitions/Menu"
            }
          },
          "400": {
            "description": "Failed"
          }
        }
      }
    },
...
    "/chef/login": {
      "post": {
        "tags": [
          "chef"
        ],
        "summary": "For chef to login to app",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful",
            "schema": {
              "$ref": "#/definitions/Chef"
            }
          },
          "401": {
            "description": "Unanthorized"
          }
        }
      }
    },
    "/chef/register": {
      "post": {
        "tags": [
          "chef"
        ],
        "summary": "For chef to register account",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "emailAddress": {
                  "type": "string"
                },
                "phoneNumber": {
                  "type": "string"
                },
                "location": {
                  "type": "object",
                  "properties": {
                    "address": {
                      "type": "string"
                    },
                    "longitude": {
                      "type": "number"
                    },
                    "latitude": {
                      "type": "number"
                    }
                  }
                },
                "storehours": {
                  "type": "object",
                  "properties": {
                    "open": {
                      "type": "number"
                    },
                    "close": {
                      "type": "number"
                    },
                    "saturday": {
                      "type": "boolean"
                    },
                    "sunday": {
                      "type": "boolean"
                    }
                  }
                },
                "profilePhoto": {
                  "type": "string"
                },
                "licencePhoto": {
                  "type": "string"
                },
                "cuisineType": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful"
          },
          "401": {
            "description": "Failed"
          }
        }
      }
    },
    "/chef": {
      "put": {
        "tags": [
          "chef"
        ],
        "summary": "Edit chef information",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "type": "object",
              "properties": {
                "password": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "emailAddress": {
                  "type": "string"
                },
                "phoneNumber": {
                  "type": "string"
                },
                "location": {
                  "type": "object",
                  "properties": {
                    "address": {
                      "type": "string"
                    },
                    "longitude": {
                      "type": "number"
                    },
                    "latitude": {
                      "type": "number"
                    }
                  }
                },
                "storehours": {
                  "type": "object",
                  "properties": {
                    "open": {
                      "type": "number"
                    },
                    "close": {
                      "type": "number"
                    },
                    "saturday": {
                      "type": "boolean"
                    },
                    "sunday": {
                      "type": "boolean"
                    }
                  }
                },
                "profilePhoto": {
                  "type": "string"
                },
                "licencePhoto": {
                  "type": "string"
                },
                "cuisineType": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful",
            "schema": {
              "$ref": "#/definitions/Chef"
            }
          },
          "401": {
            "description": "Failed"
          }
        }
      }
    },
    "/chef/menus": {
      "post": {
        "tags": [
          "chef"
        ],
        "summary": "Create new menu for chef",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Menu"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful",
            "schema": {
              "$ref": "#/definitions/Menu"
            }
          },
          "401": {
            "description": "Failed"
          }
        }
      }
    },
    "/chef/menus/:id/dishs": {
      "post": {
        "tags": [
          "chef"
        ],
        "summary": "Create new dish for specified chef menu",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Dish"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful",
            "schema": {
              "$ref": "#/definitions/Dish"
            }
          },
          "400": {
            "description": "Failed"
          }
        }
      }
    },
    "/chef/orders": {
      "get": {
        "tags": [
          "chef"
        ],
        "summary": "Get orders for chef",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Successful",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Order"
              }
            }
          },
          "400": {
            "description": "Failed"
          }
        }
      }
    },
    "/chef/orders/{id}": {
      "put": {
        "tags": [
          "chef"
        ],
        "summary": "Update order status",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "type": "object",
              "properties": {
                "accepted": {
                  "type": "boolean"
                },
                "status": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful",
            "schema": {
              "$ref": "#/definitions/Order"
            }
          },
          "400": {
            "description": "Failed"
          }
        }
      }
    },
    "/chef/summary": {
      "get": {
        "tags": [
          "chef"
        ],
        "summary": "Get chef summary",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Successful",
            "schema": {
              "type": "object",
              "properties": {
                "text": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Failed"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Account": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "activated": {
          "type": "boolean"
        }
      }
    },
...
    "User": {
      "type": "object",
      "properties": {
        "account": {
          "$ref": "#/definitions/Account"
        },
        "name": {
          "type": "string"
        },
        "emailAddress": {
          "type": "string"
        },
        "phoneNumber": {
          "type": "string"
        },
        "homeLocation": {
          "type": "object",
          "properties": {
            "address": {
              "type": "string"
            },
            "longitude": {
              "type": "number"
            },
            "latitude": {
              "type": "number"
            }
          }
        },
        "workLocation": {
          "type": "object",
          "properties": {
            "address": {
              "type": "string"
            },
            "longitude": {
              "type": "number"
            },
            "latitude": {
              "type": "number"
            }
          }
        },
        "points": {
          "type": "number"
        }
      }
    },
    "Chef": {
      "type": "object",
      "properties": {
        "account": {
          "$ref": "#/definitions/Account"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "emailAddress": {
          "type": "string"
        },
        "phoneNumber": {
          "type": "string"
        },
        "rating": {
          "type": "object",
          "properties": {
            "stars": {
              "type": "number"
            },
            "votes": {
              "type": "number"
            }
          }
        },
        "location": {
          "type": "object",
          "properties": {
            "address": {
              "type": "string"
            },
            "longitude": {
              "type": "number"
            },
            "latitude": {
              "type": "number"
            }
          }
        },
        "storehours": {
          "type": "object",
          "properties": {
            "open": {
              "type": "number"
            },
            "close": {
              "type": "number"
            },
            "saturday": {
              "type": "boolean"
            },
            "sunday": {
              "type": "boolean"
            }
          }
        },
        "profilePhoto": {
          "type": "string"
        },
        "licencePhoto": {
          "type": "string"
        },
        "cuisineType": {
          "type": "string"
        },
        "menus": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Menu"
          }
        },
        "paymentRatio": {
          "type": "number"
        }
      }
    },
    "Menu": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "dishs": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Dish"
          }
        }
      }
    },
    "Dish": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "price": {
          "type": "number"
        },
        "photo": {
          "type": "string"
        },
        "cuisineType": {
          "type": "string"
        },
        "cooktime": {
          "type": "number"
        }
      }
    },
    "Order": {
      "type": "object",
      "properties": {
        "user": {
          "$ref": "#/definitions/User"
        },
        "chef": {
          "$ref": "#/definitions/Chef"
        },
        "dishs": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Dish"
          }
        },
        "discount": {
          "$ref": "#/definitions/Discount"
        },
        "note": {
          "type": "string"
        },
        "price": {
          "type": "number"
        },
        "status": {
          "type": "string"
        },
        "accepted": {
          "type": "boolean"
        },
        "deliveryLocation": {
          "type": "object",
          "properties": {
            "address": {
              "type": "string"
            },
            "longitude": {
              "type": "number"
            },
            "latitude": {
              "type": "number"
            }
          }
        }
      }
    },
    "Discount": {
      "type": "object",
      "properties": {
        "codestring": {
          "type": "string"
        },
        "discountRate": {
          "type": "number"
        },
        "beenUsed": {
          "type": "boolean"
        }
      }
    }
  }
}